# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Anthology API
  description: |
    ## Overview

    The Anthology API provides programmatic access to oral history conversations, audio recordings, and their
    visual representation data. This API powers the Anthology visualization application, which displays
    interconnected conversations as an interactive graph where responses, questions, narratives, and speakers
    form a web of relationships.

    ### Key Concepts

    - **Anthology**: A collection of related conversations, typically from a specific project, community, or theme
    - **Recording**: An audio file associated with an anthology, with metadata such as duration, format, and file size
    - **Conversation**: A recorded discussion containing multiple responses, questions, and narratives
    - **Response**: A segment of speech from a single speaker, the core content unit
    - **Question**: Facilitator prompts that elicit responses from participants
    - **Narrative**: Thematic groupings that connect related responses across a conversation
    - **Speaker**: A participant in a conversation with associated visual styling

    ### Data Relationships

    ```
    Anthology
      ├── Recordings (1:many)
      └── Conversations (1:many)
            ├── Speakers (1:many)
            ├── Questions (1:many)
            │     └── Responses (1:many via respondsToQuestionId)
            ├── Narratives (1:many)
            │     └── Responses (1:many via respondsToNarrativeId)
...
    ```

...

...

...

...

...
servers:
...
paths:
  /anthologies:
    get:
      summary: List anthologies
      description: |
        Retrieve a paginated list of anthologies. By default, only public anthologies are returned.
        Use this endpoint to populate anthology selection interfaces or discover available content.

        **Example use cases:**
        - Landing page showing all public anthologies
        - Admin dashboard listing all anthologies (with `publicOnly=false`)

        **Note:** Private anthologies require authentication and appropriate permissions.
      operationId: listAnthologies
      tags:
        - Anthologies
      parameters:
        - name: publicOnly
          in: query
          description: |
            When `true` (default), only returns anthologies marked as public.
            Set to `false` to include private anthologies (requires authentication).
          schema:
            type: boolean
            default: true
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of anthologies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnthologyListResponse'
              example:
                data:
                  - id: "550e8400-e29b-41d4-a716-446655440000"
                    slug: "community-voices-2024"
                    title: "Community Voices 2024"
                    description: "Oral histories from local community members"
                    isPublic: true
                    createdAt: "2024-01-15T10:30:00Z"
                meta:
                  total: 12
                  limit: 50
                  offset: 0
                  hasMore: false
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /anthologies/{slug}:
    get:
      summary: Get anthology by slug
      description: |
        Retrieve a single anthology by its URL-friendly slug. Slugs are unique identifiers
        suitable for use in URLs (e.g., `/anthology/community-voices-2024`).

        **When to use:**
        - Deep linking to a specific anthology
        - Resolving anthology data from a URL route

        **Note:** Returns 404 if the anthology doesn't exist or if it's private and the user lacks access.
      operationId: getAnthologyBySlug
      tags:
        - Anthologies
      parameters:
        - name: slug
          in: path
          required: true
          description: URL-friendly unique identifier for the anthology (e.g., "community-voices-2024")
          schema:
            type: string
          example: "community-voices-2024"
      responses:
        '200':
          description: Anthology details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnthologyResponse'
              example:
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  slug: "community-voices-2024"
                  title: "Community Voices 2024"
                  description: "Oral histories from local community members"
                  isPublic: true
                  createdAt: "2024-01-15T10:30:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /recordings:
    get:
      summary: List recordings
      description: |
        Retrieve a paginated list of audio recordings. Optionally filter by anthology
        to get all recordings within a specific collection.

        **Filtering:**
        - `anthologyId`: Get recordings from a specific anthology

        **Use cases:**
        - List all audio files associated with an anthology
        - Browse recordings for playback or management
      operationId: listRecordings
      tags:
        - Recordings
      parameters:
        - name: anthologyId
          in: query
          description: Filter recordings belonging to a specific anthology
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of recordings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /conversations:
    get:
      summary: List conversations
      description: |
        Retrieve a paginated list of conversations. Optionally filter by anthology to get
        all conversations within a specific collection.

        **Example use cases:**
        - List all conversations in an anthology for navigation
        - Browse conversations across all anthologies

        **Performance note:** For visualization loading, prefer `/graph/load` which returns
        all related data in a single optimized request.
      operationId: listConversations
      tags:
        - Conversations
      parameters:
        - name: anthologyId
          in: query
          description: Filter conversations belonging to a specific anthology
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of conversations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /conversations/{id}:
    get:
      summary: Get conversation by ID
      description: |
        Retrieve detailed information about a single conversation including its metadata,
        audio file reference, and duration.

        The `metadata` object contains rich contextual information:
        - `title`: Display name for the conversation
        - `date`: When the conversation was recorded
        - `participants`: List of participant names
        - `speakerColors`: Color assignments for visualization
        - `location`: Where the conversation took place
        - `facilitator`: Who led the conversation
        - `topics`: Subject tags for categorization

        **Related endpoints:**
        - Use `/conversations/{id}/speakers` for speaker details
        - Use `/conversations/{id}/responses` for full transcript
      operationId: getConversationById
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier for the conversation
          schema:
            type: string
            format: uuid
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
      responses:
        '200':
          description: Conversation details with metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /conversations/{id}/speakers:
    get:
      summary: Get speakers for a conversation
      description: |
        Retrieve all speakers participating in a specific conversation. Each speaker includes
        their complete color scheme for consistent visualization rendering.

        **Color scheme fields:**
        - Primary colors (`circleColor`, `quoteRectangleColor`, `quoteTextColor`) for focused state
        - Faded colors for unfocused/background state

        **Ordering:** Speakers are returned in creation order, typically reflecting their
        first appearance in the conversation.
      operationId: getConversationSpeakers
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of speakers in the conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpeakerListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /conversations/{id}/questions:
    get:
      summary: Get questions for a conversation
      description: |
        Retrieve all facilitator questions from a specific conversation. Questions serve as
        prompts that elicit responses from participants.

        **Graph relationships:** Each question has a `relatedResponses` array containing
        IDs of responses that directly answer this question. These form edges in the
        visualization graph.

        **Audio timing:** `audioStartMs` and `audioEndMs` enable playback synchronization
        with the conversation recording.
      operationId: getConversationQuestions
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of questions in the conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /conversations/{id}/responses:
    get:
      summary: Get responses for a conversation
      description: |
        Retrieve all responses (speech segments) from a specific conversation. This is the
        primary content of the conversation, containing the full transcript segmented by speaker.

        **Ordering:** Responses are returned in chronological order based on `audioStartMs`
        or `turnNumber`, representing the natural flow of the conversation.

        **Performance note:** Responses may include large `embedding` arrays (1536 dimensions)
        and `wordTimestamps` arrays. For visualization-only use cases, consider using
        `/graph/load` which optimizes the data payload.
      operationId: getConversationResponses
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of responses in the conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /conversations/{id}/narratives:
    get:
      summary: Get narratives for a conversation
      description: |
        Retrieve all narratives (thematic groupings) for a specific conversation. Narratives
        are curator-created labels that connect related responses around common themes.

        **Graph visualization:** Narratives appear as colored nodes that cluster responses
        sharing a theme. The `relatedResponses` array identifies which responses belong
        to each narrative.

        **Example narratives:**
        - "Childhood memories"
        - "Career transitions"
        - "Community impact"
      operationId: getConversationNarratives
      tags:
        - Conversations
      parameters:
        - name: id
          in: path
          required: true
          description: Conversation UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of narratives in the conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NarrativeListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /questions:
    get:
      summary: List questions
      description: |
        Retrieve a paginated list of facilitator questions. Filter by conversation or anthology
        to narrow results.

        **Filtering:**
        - `conversationId`: Get questions from a specific conversation
        - `anthologyId`: Get questions from all conversations in an anthology

        **Use cases:**
        - Building a question index across an anthology
        - Analyzing facilitator patterns across conversations
      operationId: listQuestions
      tags:
        - Questions
      parameters:
        - name: conversationId
          in: query
          description: Filter questions belonging to a specific conversation
          schema:
            type: string
            format: uuid
        - name: anthologyId
          in: query
          description: Filter questions belonging to any conversation in this anthology
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionPaginatedResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /questions/{id}/responses:
    get:
      summary: Get responses for a question
      description: |
        Retrieve all responses that directly answer a specific question. This returns responses
        where `respondsToQuestionId` matches the given question ID.

        **Graph relationship:** This endpoint returns the outgoing edges from a question node
        to its connected response nodes in the visualization.

        **Ordering:** Responses are returned in conversation order (by `turnNumber` or `audioStartMs`),
        showing how participants answered the question in sequence.
      operationId: getQuestionResponses
      tags:
        - Questions
      parameters:
        - name: id
          in: path
          required: true
          description: Question UUID to find responses for
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of responses answering this question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /responses:
    get:
      summary: List responses
      description: |
        Retrieve a paginated list of responses with flexible filtering options.
        Responses are the core content units containing speaker utterances.

        **Filtering (can be combined):**
        - `conversationId`: Responses from a specific conversation
        - `anthologyId`: Responses from all conversations in an anthology
        - `questionId`: Responses answering a specific question
        - `narrativeId`: Responses belonging to a narrative theme

        **Performance note:** Responses may include large `embedding` arrays.
        For bulk loading, use `/graph/load` which optimizes data transfer.
      operationId: listResponses
      tags:
        - Responses
      parameters:
        - name: conversationId
          in: query
          description: Filter by parent conversation
          schema:
            type: string
            format: uuid
        - name: anthologyId
          in: query
          description: Filter by parent anthology (includes all conversations)
          schema:
            type: string
            format: uuid
        - name: questionId
          in: query
          description: Filter by responses that answer this question
          schema:
            type: string
            format: uuid
        - name: narrativeId
          in: query
          description: Filter by responses belonging to this narrative
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of responses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponsePaginatedResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create a response
      description: |
        Create a new response within a conversation. Requires authentication.

        **Required fields:**
        - `conversationId`: Parent conversation
        - `speakerName`: Who is speaking
        - `speakerText`: Full transcript of the utterance

        **Optional relationships:**
        - `respondsToQuestionId`: Links response to a facilitator question
        - `respondsToResponseId`: Links response to another participant's response
        - `respondsToNarrativeId`: Links response to a thematic narrative

        **Audio synchronization:**
        - `audioStartMs` / `audioEndMs`: Timing for playback sync
        - Word timestamps are generated separately via processing pipeline

        **Note:** Embeddings are generated server-side and cannot be provided directly.
      operationId: createResponse
      tags:
        - Responses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResponseRequest'
            example:
              conversationId: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
              speakerName: "Maria"
              speakerText: "When I think about my childhood, I remember the summers at my grandmother's house."
              pullQuote: "I remember the summers at my grandmother's house"
              respondsToQuestionId: "8f14e45f-ceea-467f-a10a-6d6c3b76b98e"
              audioStartMs: 125000
              audioEndMs: 132500
              turnNumber: 5
              medium: "audio"
              synchronicity: "sync"
      responses:
        '201':
          description: Response created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /responses/{id}:
    get:
      summary: Get response by ID
      description: |
        Retrieve a single response with all its details including transcript,
        timing information, relationships, and word timestamps.

        **Included data:**
        - Full speaker text and optional pull quote
        - Audio timing (`audioStartMs`, `audioEndMs`)
        - Relationship links (`respondsToQuestionId`, etc.)
        - Word-level timestamps for karaoke highlighting
        - Embedding vector (1536 dimensions) for similarity search
      operationId: getResponseById
      tags:
        - Responses
      parameters:
        - name: id
          in: path
          required: true
          description: Response UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Response details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Update a response
      description: |
        Update specific fields of an existing response. Requires authentication.

        **Updatable fields:**
        - `speakerText`: Correct transcript errors
        - `pullQuote`: Set/update highlighted excerpt
        - `respondsToNarrativeId`: Change narrative association
        - `audioStartMs` / `audioEndMs`: Adjust timing

        **Non-updatable fields:**
        - `conversationId`, `speakerName`: Cannot be changed after creation
        - `embedding`: Regenerated automatically when `speakerText` changes
        - `wordTimestamps`: Managed separately via processing pipeline

        **Partial updates:** Only include fields you want to change.
      operationId: updateResponse
      tags:
        - Responses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Response UUID to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResponseRequest'
            example:
              pullQuote: "summers at my grandmother's house"
              audioStartMs: 125500
              audioEndMs: 132000
      responses:
        '200':
          description: Response updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete a response
      description: |
        Permanently delete a response. Requires authentication.

        **Warning:** This action cannot be undone. Deleting a response will:
        - Remove the response from its conversation
        - Break any links from other responses (`respondsToResponseId`)
        - Remove the response from associated narratives
        - Delete all associated word timestamps
      operationId: deleteResponse
      tags:
        - Responses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Response UUID to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Response deleted successfully (no content returned)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /responses/{id}/word-timestamps:
    get:
      summary: Get word timestamps for a response
      description: |
        Retrieve word-level timing data for karaoke-style highlighting during audio playback.
        Each word includes start/end times in milliseconds relative to the conversation audio.

        **Use cases:**
        - Synchronized word highlighting during audio playback
        - Jump-to-word navigation
        - Accessibility features (visual feedback for speech)

        **Data structure:**
        Each timestamp includes:
        - `text`: The word or punctuation
        - `start`: Start time in milliseconds
        - `end`: End time in milliseconds
        - `speaker`: Speaker identifier (for multi-speaker segments)
        - `confidence`: Transcription confidence score (0-1)

        **Note:** Not all responses have word timestamps. Returns empty array if unavailable.
      operationId: getResponseWordTimestamps
      tags:
        - Responses
      parameters:
        - name: id
          in: path
          required: true
          description: Response UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Word timestamps for the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WordTimestampListResponse'
              example:
                data:
                  - text: "When"
                    start: 125000
                    end: 125200
                    speaker: "Maria"
                    confidence: 0.98
                  - text: "I"
                    start: 125200
                    end: 125300
                    speaker: "Maria"
                    confidence: 0.99
                  - text: "think"
                    start: 125300
                    end: 125550
                    speaker: "Maria"
                    confidence: 0.97
        '500':
          $ref: '#/components/responses/InternalError'
  /narratives:
    get:
      summary: List narratives
      description: |
        Retrieve a paginated list of thematic narratives. Narratives are curator-created
        groupings that connect related responses around common themes or topics.

        **Filtering:**
        - `conversationId`: Narratives from a specific conversation
        - `anthologyId`: Narratives from all conversations in an anthology

        **Use cases:**
        - Display narrative themes for navigation
        - Build a theme index across an anthology
        - Analyze thematic patterns across conversations
      operationId: listNarratives
      tags:
        - Narratives
      parameters:
        - name: conversationId
          in: query
          description: Filter narratives belonging to a specific conversation
          schema:
            type: string
            format: uuid
        - name: anthologyId
          in: query
          description: Filter narratives belonging to any conversation in this anthology
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Paginated list of narratives
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NarrativePaginatedResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT token obtained through Supabase Auth.

        Include in the Authorization header:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```

        Tokens expire after the configured Supabase session duration.
        Refresh tokens before expiration to maintain access.
  parameters:
    limit:
      name: limit
      in: query
      description: |
        Maximum number of items to return per page.
        Use with `offset` for pagination.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...